+++ CHANGES.rst
@@ -14,6 +14,9 @@
 - FileStorage: fix a rare data corruption when using restore after multiple undos.
   For details see `#395 <https://github.com/zopefoundation/ZODB/pull/395>`_.
 
+- Fix exit code of ``repozo`` script in case of verification error.
+  For details see `#396 <https://github.com/zopefoundation/ZODB/pull/396>`_.
+
 
 5.8.1 (2023-07-18)
 ==================